# docs/api-openapi.yaml
openapi: 3.0.3
info:
  title: QuicK ReeLS API
  version: 0.1.0
  description: >
    Business API used by the TikTok Mini H5 and the separate operations console.
    Provider credentials and raw media URLs never leave the server.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production placeholder
...
paths:
...
  /admin/ui-components:
    get:
      tags: [Operations]
      summary: List UI component switches
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: UI component switches
  /admin/ui-components/{key}:
    patch:
      tags: [Operations]
      summary: Toggle a public UI component
      security: [{ adminBearerAuth: [] }]
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
                page: { type: string, enum: [APP, HOME, ALBUM, PROFILE] }
      responses:
        '200':
          description: Updated component
...
  /admin/auth/login:
    post:
      tags: [Operations]
      summary: Log in an operations administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Administrator session
        '401':
          description: Invalid credentials
...
  /admin/albums:
    get:
      tags: [Operations]
      summary: List albums for operations
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: Album management list
    post:
      tags: [Operations]
      summary: Create an album draft
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: Created draft
  /admin/upload-jobs:
    get:
      tags: [Operations]
      summary: List provider upload jobs
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: Upload jobs
    post:
      tags: [Operations]
      summary: Create an asynchronous upload job
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: Created upload job
  /admin/upload-jobs/{jobId}:
    get:
      tags: [Operations]
      summary: Get an upload job
      security: [{ adminBearerAuth: [] }]
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Upload job
        '404':
          description: Upload job not found
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminBearerAuth:
      type: http
      scheme: bearer